extract builtin exit to add more condition check, unset didnt finish but cd finished
