+++ include/executor.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/21 10:13:43 by spyun         #+#    #+#                 */
-/*   Updated: 2025/01/21 22:15:53 by bewong        ########   odam.nl         */
+/*   Updated: 2025/01/21 23:13:44 by bewong        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,8 @@
 # include "minishell.h"
 
 # define PATH_MAX 1024
+# define MAX_STATUS_LEN 20
+# define EXIT_ERROR_CODE 255
 
 typedef struct s_exec
 {
